Rename the Mage zero install option file and add ignore build tag
The mage zero install option file is used for more easy builds.
I this case, I changed the file name from 'main.go' to 'mage.go'
because it's more clear what it's related to. And since it's not
allowed to have multiple packages in a single directory in Go, I've
added the 'ignore' build tag that makes it an exception and would
be allowed.